parser/htmlparser/tests/mochitest/test_bug667533.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=667533
-->
<head>
  <title>Test for Bug 667533</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=667533">Mozilla Bug 667533</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

function loaded(iframe) {
  is(iframe.contentWindow.location.href, iframe.src, "should load correct URL");
  is(iframe.contentDocument.body.textContent, '{"<p>Hello</p>": null}', "application/json should be treated as text");
  SimpleTest.finish();
}
</script>
<iframe src="data:application/json,{&quot;<p>Hello</p>&quot;:%20null}" onload="loaded(this);"></iframe>
</body>
</html>
